+++ warthog/__init__.py
@@ -11,6 +11,7 @@
 warthog
 ~~~~~~~
 
+Simple client for A10 load balancers.
 """
 
-__version__ = '0.7.0'
+__version__ = '0.8.0'