+++ warthog/exceptions.py
@@ -33,6 +33,58 @@ class WarthogError(Exception):
         return '{clazz}("{msg}")'.format(clazz=self.__class__.__name__, msg=self.msg)
 
 
+class WarthogConfigError(WarthogError):
+    """Base for errors raised while parsing or loading configuration."""
+
+
+class WarthogNoConfigFileError(WarthogConfigError):
+    """No configuration file could be found."""
+
+    def __init__(self, msg, locations_checked=None):
+        """Set the message for this exception and an optional list of paths
+        that were checked for configuration files.
+        """
+        super(WarthogNoConfigFileError, self).__init__(msg)
+        self.locations_checked = list(locations_checked) if locations_checked is not None else []
+
+    def __str__(self):
+        out = [self.msg]
+        if self.locations_checked is not None:
+            out.append('Locations checked: ' + ', '.join(self.locations_checked))
+        return '. '.join(out)
+
+    def __repr__(self):
+        return '{clazz}("{msg}", locations_checked={locations_checked})'.format(
+            self.msg, self.locations_checked)
+
+
+class WarthogMalformedConfigFileError(WarthogConfigError):
+    """The configuration file is missing required sections or fields."""
+
+    def __init__(self, msg, missing_section=None, missing_option=None):
+        """Set the message for this exception and optional missing sections
+        or options expected in the configuration file.
+        """
+        super(WarthogMalformedConfigFileError, self).__init__(msg)
+        self.missing_section = missing_section
+        self.missing_option = missing_option
+
+    def __str__(self):
+        out = [self.msg]
+        if self.missing_section is not None:
+            out.append('Missing-section: {0}'.format(self.missing_section))
+        if self.missing_option is not None:
+            out.append('Missing-option: {0}'.format(self.missing_option))
+        return '. '.join(out)
+
+    def __repr__(self):
+        return ('{clazz}("{msg}", missing_section="{missing_section}", '
+                'missing_option="{missing_option}")').format(clazz=self.__class__.__name__,
+                                                             msg=self.msg,
+                                                             missing_section=self.missing_section,
+                                                             missing_option=self.missing_option)
+
+
 class WarthogApiError(WarthogError):
     """Base for errors raised in the course of interacting with the load balancer.